+++ packages/sources/tiingo/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## v1.1.0
 
+## 1.4.0
+
+### Minor Changes
+
+- cd01b0c5: Change crypto WS endpoint and fix incorrect WS cache key usage
+
+## 1.3.0
+
+### Minor Changes
+
+- 56be0406: fix ws
+
 ## 1.2.10
 
 ### Patch Changes